refactor: place hooks in separate folder
